Move constants/layers.ts -> constants/map/layerStyles.ts
