+++ app/src/app/components/ContextMenu.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {ContextMenu, Text} from '@radix-ui/themes';
 import {useMapStore} from '@/app/store/mapStore';
 import {useAssignmentsStore} from '@/app/store/assignmentsStore';
-import {CHILD_LAYERS, PARENT_LAYERS} from '../constants/layers';
+import {CHILD_LAYERS, PARENT_LAYERS} from '../constants/map/layerStyle';
 
 export const MapContextMenu: React.FC = () => {
   const mapDocument = useMapStore(state => state.mapDocument);
