+++ app/src/app/components/Map/PointLayers/PointSelectionLayer.tsx
@@ -3,7 +3,7 @@ import {
   BLOCK_POINTS_LAYER_ID_CHILD,
   SELECTION_POINTS_SOURCE_ID,
   SELECTION_POINTS_SOURCE_ID_CHILD,
-} from '@/app/constants/layers';
+} from '@/app/constants/map/layerStyle';
 import {useLayerFilter} from '@/app/hooks/useLayerFilter';
 import React from 'react';
 import {Layer} from 'react-map-gl/maplibre';
