+++ app/src/app/components/Map/PolygonLayers/DemographicLayer.tsx
@@ -1,7 +1,7 @@
 import type React from 'react';
 import {
   BLOCK_SOURCE_ID,
-} from '@/app/constants/layers';
+} from '@/app/constants/map/layerStyle';
 import {useMapControlsStore} from '@/app/store/mapControlsStore';
 import type {FilterSpecification} from 'maplibre-gl';
 import {Layer} from 'react-map-gl/maplibre';
