+++ app/src/app/components/Map/PolygonLayers/GeometryOutlineLayer.tsx
@@ -1,5 +1,5 @@
 import type React from 'react';
-import {BLOCK_SOURCE_ID} from '@/app/constants/layers';
+import {BLOCK_SOURCE_ID} from '@/app/constants/map/layerStyle';
 import {Layer} from 'react-map-gl/maplibre';
 import {FilterSpecification} from 'maplibre-gl';
