+++ app/src/app/components/Map/PolygonLayers/ZoneLayers/ZoneHighlightLayer.tsx
@@ -1,5 +1,5 @@
 import type React from 'react';
-import {BLOCK_SOURCE_ID} from '@/app/constants/layers';
+import {BLOCK_SOURCE_ID} from '@/app/constants/map/layerStyle';
 import {useMapControlsStore} from '@/app/store/mapControlsStore';
 import type {FilterSpecification} from 'maplibre-gl';
 import {Layer} from 'react-map-gl/maplibre';
