+++ app/src/app/components/Toolbar/ColorPicker.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useRef} from 'react';
 import {useMapStore} from '@/app/store/mapStore';
 import {NullableZone} from '@/app/constants/types';
-import {FALLBACK_NUM_DISTRICTS} from '@/app/constants/layers';
+import {FALLBACK_NUM_DISTRICTS} from '@/app/constants/map/layerStyle';
 import {ColorRadioGroup} from './ColorRadioGroup';
 import {ColorDropdown} from './ColorDropdown';
 import {ColorMultiDropdown} from './ColorMultiDropdown';
