+++ app/src/app/constants/map/layerStyle.ts
