+++ app/src/app/utils/events/mapEvents.ts
@@ -24,7 +24,7 @@ import {
   BLOCK_POINTS_LAYER_ID_CHILD,
   BLOCK_SOURCE_ID,
   INTERACTIVE_LAYERS,
-} from '@/app/constants/layers';
+} from '@/app/constants/map/layerStyle';
 import {ResetMapSelectState} from '@utils/events/handlers';
 import GeometryWorker from '../GeometryWorker';
 import {ActiveTool} from '@/app/constants/types';
